requirements.txt
FLASK==1.1.1
werkzeug==0.16.0

# Application server for both development and production.
gunicorn==20.0.4

# Testing and static analysis.
pytest==5.4.2
pytest-cov==2.8.1
flake8==3.7.9

# Data
psycopg2==2.8.5
Flask-SQLAlchemy==2.4.1
sqlalchemy-utils==0.36.5

# Forms.
Flask-WTF==0.14.3
WTForms-Components==0.10.4

# Extensions.
flask-debugtoolbar==0.10.1
Flask-Static-Digest==0.1.3
Flask-Login==0.5.0

#faker
Faker==4.1.0